game_development
By Karla Garcia Corral Lobo
January 22, 2023
Hult International Business School

This assignment is for the Python for Data Analysts: Methods & Tools class with Professor Chase Kusterer. The goal is to build a text adventure game based on a movie or book, I chose to base this game on the film Toy Story. 

Enjoy the game!