Classification Model Development
This assignment focuses on analyzing Apprentice Chef’s customer database to determine the variables that have the most impact on the cross-selling success of the business. The “CROSS_SELL_SUCCESS” will serve as the dependent variable (y), while the other variables are used are the interdepend variables (x). This database contains approximately 2,000 customers who made at least one purchase per month for 11 months in their first year, at least one purchase per quarter, or a minimum of 15 purchases throughout their first year.

The goal was to identify the variables that have the most significant impact on cross-selling, which can help Apprentice Chef optimize its sales efforts and increase its bottom line. In addition, new features were developed to improve the interpretability of our data and the performance of the model.

This study begins by examining Chef Apprentice's customer data to gain an understanding of consumer behavior patterns and their impact on the business's revenue. In this analysis, we first determined the variables that needed to be cleaned and identified any new potential features that could be created based on the existing information in the customer database.