Regression Model Development
This assignment focuses on analyzing Apprentice Chef’s customer database to determine the variables that have the most impact on the revenue of the business. The revenue will serve as the dependent variable (y), while the other variables are used as the independent variables (x). This database contains approximately 2,000 customers who made at least one purchase per month for 11 months in their first year, at least one purchase per quarter, or a minimum of 15 purchases throughout their first year.

Our objective was to identify the variables that have the most significant impact on revenue, which can help Apprentice Chef optimize its sales efforts and increase its bottom line. In addition, we developed new features to improve the interpretability of our data and the performance of the model.

This study begins by examining Chef Apprentice's customer data to gain an understanding of consumer behavior patterns and their impact on the business's revenue. In this analysis, we first determined the variables that needed to be cleaned and identified any new potential features that could be created based on the existing information in the customer database.